series: Create Series objects when parsing mails
This commit only adds basic support for the initial series submission.

Series with or without cover letters are supported. When sent without a
cover letter, the series is named "Untitled series". It's planned to let
the user chose a more appropriate series title through the web UI at a
later point.

Single patches are treated as a Series of 1 patch, named with the
subject of that patch.

v2: SERIES_DEFAULT_NAME change

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>
Acked-by: Stephen Finucane <stephen.finucane@intel.com>