testresult: Add Test and TestResult models
The idea is to store test results in patchwork itself so the series and
patch pages can show a summary of the testing done.

Tests are defined per-project, so they can be administrated at the
project level (eg. configure the mailing policy).

We only store one test results per test, the latest state updated by the
testing system.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>