tests: Use --noinput when running tests
We don't want prompts when running tests, especially when they are run
through tox and most likely on a CI server.

When cancelling a test mid-way, the test db is not destroyed ans
subsequent runs will fail trying to create it and asking the user what
to do. --noinput will use the default answer, that is to destroy the
test db and start anew.

v2: Rebase on top of upstream
v3: Rebase, yet again.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>
Acked-by: Stephen Finucane <stephen.finucane@intel.com> (v2)