project: Provide a setting to only consider git send-email patches
I've seen quite a few false positive, ie. mail with diffs in them
(because people use diffs to illustrate an idea) were treated as
patches.

One goal I have is to lower patchwork maintenance (ie limit the number
of fixes we need to do by hand). That's a small improvement towards that
direction.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>