api: Document and unit test how list of objects are exposed
We'll describe all corners of the REST API, little by little. This time,
we're talking about how list of objects work. More importantly, a unit
test is in place to ensure what's documented doesn't regress.

Technically, we're changing how the list of revisions is exposed. It
goes from a JSON array to a full blow paginated list. It's probably a
bit overkill for the list of revisions, by I have seen stupid number of
revisions in the past and with the expand GET parameter, such number
could a be problem.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>