api: Correctly order patches in revisions
One of those cases where frameworks don't really help.

1/ Django doesn't respect the ordering defined in 'through' models of
ManytoMany field when accessed using the ManyRelatedManager.

2/ Rest framework don't really let you define the specifics of a
queryset for a related field easily

On top of that, I added the related field expansion through a GET
parameter, so we still need to take that into account.

The solution found is to do the work ourselves, binding a
SerializerMethodField to the 'patches' related field and write the full
logic.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>