api: Make all DateTimeFields serialize time in ISO 8601 format
It'd be handy to be able to reinject the content of the DateTimeFields
from the JSON response into the API queries. We'll need that in a
'since' GET parameter for the list of events.

However, datetime objects in python have more precision that the
millisecond currently exposed in the JSON API, becaure rest frameworks
defaults to supporting dates as described in ECMA 262. This mistmach
causes django queries to be wong, because we are dealing with a
truncated time, so operators like gt and gte don't properly
exclude/include the query timesmap.

This also means we'll have to be careful when parsing the date in
javascript later on, not all browser support full ISO 8601 dates (IE,
I'm looking at you). I was already planning to use momentjs, that's
another good reason to do so.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>