tests: Don't use rest_framework APITest client
APITest was lying to us!

The JSON data retrieved from APITest's get().data wasn't the actual data
parsed from the HTTP response but the internal structure of the
Serializer. This means we were having datatime objects in the dict that
was supposed to be decoded from the JSON response. In particular, those
objects had the full timestamp, including the microseconds, which isn't
true when actually decoding the data from the HTTP response, because
rest_framework uses ECMA 262 time then (truncating microseconds).

Actually decode the HTTP response so we can properly test the API.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>