api: 'related=expand' shouldn't make the number of queries explode
While related=expand isn't officially part of the API (it isn't
documented), it's used by the page presenting the list of series to
retrieve all the information to display in one API call.

No select_related() was done in that case, we can do better: when
related=expand is given, make sure to include the related fields in the
query.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>