tests: Introduce unit tests for the javascript side of things
Patchwork frontend needs a lot of work and, beside the seleniums tests
that are end-to-end, integration tests, there's room for proper JS unit
tests.

I've done that with karma + jasmine, it's fairly straightfoward and
actually quite nice to use. I'm a big fan of karma's "rerun the tests on
save" feature that made the development of patch_strip_series_markers()
much more fun.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>