project: Allow multiple projects to share the same mailing list
I have the case where contributions for several git repositories are
sent to the same mailing-list. We really want to handle them separately
though, some of the repos have different maintainers for instance.

An easy way to solve that is to ask people to add a special tag in
subject prefix of patches. It's can even be sticky to a specific git
checkout with:

  $ git config format.subjectprefix "PATCH evdev"

So people don't have to remember to tag their submissions by hand.

Those tags are stripped out of the patch name because they are redundant
information with the project they are now in.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>