parsemail: Ensure that the patch order cannot be None
find_patch_order() was starting with order = None and trying to retrieve
the order from the previous revision. That doesn't work when there's no
previous revision, because patchwork didn't know about series when the
previous patch was sent.

The test case triggering the failure is actually a bit involved. It
needs series as reply to a couple of revision of a single patch, with
the part of thread before series support. The test case has the full
story.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>