tools: Remove bashisms from the quick start devel scripts
Michael Ellerman noticed we weren't using shebangs in those scripts. I
shrugged it off too quickly in the review, because it worked after all.
Turned out to be a bashmism, along with the usage of 'source'.

With those fixed, we have a better chance to run on most shells (famous
last words). Tested it with bash and dash.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>