parsemail: Never drop patches sent as attachments
Even with git_send_email_only set to True, we don't want to drop patches
sent as attachments.

The purpose of git_send_email_only was to reduce the number of false
positives with people sending diff hunks inlined, but an attached patch
is always something we want to track.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>