git-pw: Add a list command
The command lets us list the series which will be displayed on stdout
either JSON or a more friendly format, the latter being the default.

The '--fields' command allows to select fields, that for the series,
these are the possibilities: 'id', 'project', 'name', 'n_patches',
'submitter', 'submitted', 'last_updated', 'version' and 'reviewer'.
Multiple fields can be selected separated by a space'

Examples:

- Print series
  $ git pw list --since 2015-12-09T13:56:21.843098

- Print series in JSON format
  $ git pw list --since 2015-12-09T13:56:21.843098 --json

- Print series, just id and series' name
  $ git pw list --since 2015-12-09T13:56:21.843098 --fields 'id name'

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>