testresult: Add a 'Reply-To' header to the mailing list in test emails
People noticed that replying to test result emails was populating the
To: field with the patchwork email address. That's not the most useful
value to have there, so default to replying to the mailing-list address.

One could make that field configurable (ie. adding it to the the Test
model) if needed.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>