api: Tune the user serializer to look like the submitter one
Making User look like Submitter has some interesting properties like
reusing the same code to display both kind of fields.

This serializer isn't part of the documented API and isn't currently
reachable anyway (because it serializes the reviewer field which can't
be set). All that to say it's totally fine to change it.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>