js: Make the initial filter clearing part of the common code
The common filter setup code is now using clear_filter() to offer a
chance to reset the filter state as part of the initialisation.
clear_filter() is now used to reset the internal filter state at both
init() time and when the filter is being cleared after having been
applied.

This gives the opportunity to go and also reset the completion state of
the submitter filter, now that clear_filter() is called after init().

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>