series: Don't allow to sort by submitter/reviewer
Allowing that is not very useful. The only real reason to sort by those
fields is to figure out all patches submitted/to review of an
individual, something that is much more convenient to do with filters.

It also make the interaction between sorting and updates much simpler,
removing so edge cases we don't really want to look at.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>