tests: Test we create events when new series change the number of patches
We had a bug where re-submitting a series with a different number of
patches wasn't correctly updated the number of patches of that series
and thus patchwork was never able to understand that the new series was
complete.

This had the unfortunate effect that CI systems listening for
"series-new-revision" events weren't ever able to see, and so test, the
new revision.

Since I moved the number of patches on the revision, this should now
work. Added a check on the number of patches and events to have some
coverage of that fix.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>