js: Also "disable" the label of disabled radio buttons
We can't actually match the parent of an element in CSS, so we need a
bit of logic to set/unset the disabled class of the label corresponding
to the radio button we want to disable.

To have that logic only in one place, I introduced a small mixin that
extends both filters and actions.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>