mail: Send a notification emails to reviewers
We're going fancy.

Because sending emails can take some time, especially if we set the
reviewer of 100 series from the web page, the mailing is done in celery
tasks.

I chose redis as the message broker as it can be used for other things
as well, like a memcache-like cache if we ever want to go that route. It
can also be used for communication with a daemon that would support long
lived connections (think web socket or server sent events)

Also, to try and be a bit generic, emails are rendered from templates.

Unit tests cover everything I could think about, including not notifying
users that have set the reviewer to themselves.

Fixes: https://github.com/dlespiau/patchwork/issues/28
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>