xmlrpc: Override the request saved by the threadlocal middleware
The auth mechanism of the XML-RPC view doesn't interact like the session
middleware: By the time the request hit the threadlocal middleware
request.user isn't populated.

So, we override the threadlocal request (and thus request.user) as soon
as we authenticate the user.

As always, if a bug manage to slip through the cracks, the punishment is
to write a test that covers it.

Fixes: https://github.com/dlespiau/patchwork/issues/168
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>