settings: Use explicit setup for the Debug Toolbar
The 'django-debug-toolbar' application provides an automatic method of
configuring the plugin. However, as noted in the documentation [1],
this can cause issues like circular imports. In this case, a
"Table 'patchwork.patchwork_state' doesn't exist" exception was being
raised when attempting to do an initial migration.

Resolve this by using the manual configuration provided in the docs.

[1] https://django-debug-toolbar.readthedocs.org/en/1.4/installation.html

v2: Add django-debug-toolbar in the requirements (Damien)
v3: Remove unused import (Damien)

Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Closes-bug: #29
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>