parsemail.sh: Always return a zero exit status
This reverts changes to parsemail.sh introduced by
cbe992d84fba57831d44afb3a21cdf83454018b2.

When parsemail is used as a delivery command from a mail server like
postfix (as it is intended to be), a non-zero exit code will cause a
bounce message to be returned to the user. From the postfix manual:

  When  the  command  fails, a limited amount of command output is
  mailed back to the  sender.   The  file  /usr/include/sysexits.h
  defines  the expected exit status codes.

For cases where patchwork is unavailable, we absolutely do not want to
start bouncing messages to all patch contributors.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Stephen Finucane <stephen.finucane@intel.com>