series: Add a new action to re-trigger a testing cycle
People are asking our CI team to re-trigger tests when something goes
wrong (for instance when a test slave was online during the previous
test run) or to double check that the results are stable.

This action allows maintainers to re-trigger test runs by inserting
series-new-revision events into the event stream.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>