parsemail: Update user's name to the last one used
Currently the first email patchwork picks up from an address defines
name assigned to that address. If it contains a typo it will stick and
require an admin to fix that.

Follow up mail with the name fixed does not change a thing.

Since the mboxes are recomposed using that name, the typo can stick to
multiple patches unnoticed, which seems to be the case quite often.

So let's fix that with updating Person with the latest name they used
from their email address.

v2: add test, and adjust existing ones for new behavior

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>