templates/patch: Link back to the series
It's hard to find the series corresponding to the patch when you have
only the patch URL - you have to go through the author or through the
project and try to find an association with the series name.

Series view is also where the results from CI are shown, so this is the
drill for many people doing the reviews.

Since this is a common issue, let's link the series from the patch page.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>