rest: Use "human name" for the series
Currently all series without a cover letter are visible both through the
API and the web interface (which uses the API) as a
"Series without cover letter".

That makes the series hard to identify and takes a toll on developers'
and maintainers' time.

We already have means of generating friendlier titles in form
"series starting with THE FIRST PATCH NAME"
that was created for seeding out results,

So let's reuse it in the API for consistency.

Also bump query number in one of the related tests, became this patch
introduces an additional one.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>