series: Improve filtering UI
Move "Filter by:" to the left, as "Status" is just a filter like the
others, so it makes sense for them to be grouped together.

So instead of:  Status     | Filter By: | Date | Submitter
We have now:    Filter By: | Status     | Date | Submitter

The "show all series" link changes its misleading name to "reset
filters", as it restores filters to their default values (i.e. does not
show series that are considered "done").

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>