series-list: Make the action buttons always visible
"Test again" and "set reviewers" buttons were hard to discover, because
as a maintainer, you had to select some checkboxes, which resulted in
filtering options being replaced with the action buttons.

Now, the action buttons are always visible but grayed out, with helpful
mouse hover tooltip explaining the requirements for them to be enabled.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>