series: Add test latest revision again button
Before, to rerun tests on a series, we had to go to series lists, search
for it, select a checkbox and press the button.

This is quality of life improvement that adds the button to the series
page, so there's less hassle required.

The button is blue, and when clicked it is disabled and turns yellow for
the duration of request to the server, then turns either green, if the
rerun was queued successfully, or red when the request has failed.

Button's text also changes to reflect the current state.

To re-enable the button (e.g. in unusual scenario of queueing two
reruns) page needs to be refreshed.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>