series: Disable "test [...] again" button when appropriate
Move the button to a div container, so it can be disabled with an
overlay with a tooltip.

The button is now disabled if the user is not logged in or is not a
maintainer of the project, with an appropriate tooltip explaining this.

Also juggle fade_{enable,disable} functions so they are exported as a
part of the pw object, so we can reuse them for disabling the
before mentioned button.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>