Stop using RequestContext and use dictionaries
RequestContext is going to be deprecated in future versions of Django in
favor of simple dictionaries.

This requires us to drop the PatchworkRequestContext class and spread
logic it contained around.

Some of the code made it into context processor but most ended up in
generic_list which actually benefits us in better code locality.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>