parsemail: Setup Django right after importing
If we won't do that importing some of the classes (e.g. User) will cause
and exception complaining about apps not being loaded with newer Django.

Because of this we also have to add an exception for module level
imports that are not on the top (i.e. before any actual code).

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>