mbox: Use the 'From' header from the incoming email
Currently all the patchwork-generated mboxes uses the name and email
address extracted from the last patch that person sent. This is strange
behaviour that leads to problem when a users puts additional information
(e.g. sponsor) in the header or has changed their legal name.

There's no silver bullet to address all the possible scenarios, and this
behaviour causes a lot of confusion.

Luckily we store the vanilla email headers the patch was sent out with
in patch.headers. We can use that to stop the guess game and let the
people to control how they are called for every patch.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>