api: Make /newrevision/ copy a revision and mark it as rerun
Up until this patch, rerunning was just creating a 'newrevision' event
for the existing revision, effectively queuing it again. Any incoming
results would overwrite the existing ones.

Now, whenever rerun is hit, the last revision will be duplicated (no
results overwriting!), bumping "Last Update" time for the series. We
also mark such series as a rerun so we will be able to display it for
the users / expose it in the API.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>