Present series state to the user
Include information that series is not complete (i.e. we have received
less patches than expected, basing on the [xx/yy] tag) and that series
looks "strange" (i.e. ordering may is off or we have got too many
patches).

Changes in logic:
If we get more patches than expected, *do not* reset status to
incomplete series, we've got the event for queued already anyway.
Just consider it "strange".

Also, let's include information which revision we are rerunning on the
rerun button.

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>